fix: process buffered output inside docker attach() (#1551)
Docker's attach() issues an HTTP POST, to which docker engine issue a
response upgrading the connection to raw TCP, over which docker's
internal protocol is used for further communication. The is processed
client-side by extracting the raw socket object from the received
HTTPResponse and reading from that.

The problem is that HTTPResponse wraps the socket with an
io.BufferedReader in order to process the response status and headers,
afer which, additional contents may still be left inside
io.BufferedReader's buffer, which will not be seen when attempting to
read from the underlying socket.

This commit ensure the io.BufferedReader's buffer gets propagated and
read before attempting further reads form the socket.